requirements.txt
Django==4.2.11
djangorestframework
django-filter
djangorestframework-simplejwt

django-allauth
dj-rest-auth

django-cors-headers
psycopg2-binary
redis
django-redis
python-dotenv
gunicorn
requests>=2.31.0
cryptography>=42.0.0

django-phonenumber-field[phonenumbers]
django-taggit
django-allauth
dj-rest-auth
djangorestframework-simplejwt
Pillow